Restart On Current Level After Dying
Instead of restarting the whole game after dying, it restarts the player on the start of the level that they died on.  It also resets their score to the baseline score for that level.  This changes the game significantly, but also gives players a chance to progress farther in the game.